Revert "Reorganized the States"
