+++ hand-painter/Dataset.py
@@ -11,7 +11,7 @@ import json
 
 class Dataset:
     def __init__(self):
-        f = open("../data/labels2.json", "r")
+        f = open("data/labels2.json", "r")
         self.labels_obj = json.load(f)
         self.labels = list(self.labels_obj.keys())
         f.close()
@@ -68,7 +68,7 @@ class Dataset:
             # canvas = cv2.imread('images/pear.png', cv2.IMREAD_GRAYSCALE)
 
             timestamp = time.time()
-            filename = f"../images/{timestamp}.png"
+            filename = f"images/{timestamp}.png"
             cv2.imwrite(filename, canvas)
 
             # Reshape
@@ -76,7 +76,7 @@ class Dataset:
             canvas = canvas.astype("float32")
 
             # Load TFLite model and allocate tensors
-            interpreter = tf.lite.Interpreter(model_path="../cnn-model/model2.tflite")
+            interpreter = tf.lite.Interpreter(model_path="cnn-model/model2.tflite")
             interpreter.allocate_tensors()
 
             # Set input tensor
