+++ hand-painter/README.md
@@ -2,7 +2,7 @@
 
 Create a virtual environment and install the dependencies:
 
-```bash
+```
 python3 -m venv ./venv
 source ./venv/bin/activate
 pip3 install -r requirements.txt
@@ -11,8 +11,8 @@ pip3 install -r requirements.txt
 
 ## To run without the "Tirar foto" button:
 
-```bash
-python3 main.py --no-photo  # inside the src/ directory
+```python
+python3 main.py --no-photo
 ```
 <br>
 
@@ -26,6 +26,6 @@ PASSWORD=very_secure_password
 
 Run the application
 
-```bash
-python3 main.py     # inside the src/ directory
+```python
+python3 main.py
 ```
