+++ hand-painter/Ranking.py
@@ -3,17 +3,16 @@ import pandas as pd
 
 class Ranking:
     top = []  # top 10 scores
-    csv_file = "../data/ranking.csv"
 
     def __init__(self):
-        df = pd.read_csv(self.csv_file)
+        df = pd.read_csv("data/ranking.csv")
         df = df.sort_values(by=["score"], ascending=False)
         df = df.reset_index(drop=True)
         # df to array of dictionaries
         self.top = df.head(10).to_dict("records")
 
     def readRanking(self):
-        df = pd.read_csv(self.csv_file)
+        df = pd.read_csv("data/ranking.csv")
         df = df.sort_values(by=["score"], ascending=False)
         df = df.reset_index(drop=True)
         self.top = df.head(10).to_dict("records")
@@ -27,10 +26,12 @@ class Ranking:
 
     def insertScore(self, name, score, draw):
         if name != "":
-            df = pd.read_csv(self.csv_file)
-            df.loc[len(df)] = [name, score, draw]
+            df = pd.read_csv("data/ranking.csv")
+            df = df.append(
+                {"name": name, "score": score, "draw": draw}, ignore_index=True
+            )
             df = df.sort_values(by=["score"], ascending=False)
             df = df.reset_index(drop=True)
-            df.to_csv(self.csv_file, index=False)
+            df.to_csv("data/ranking.csv", index=False)
             self.top = df.head(10).to_dict("records")
             print(self.top)
