+++ hand-painter/Timer.py
