+++ hand-painter/main.py
@@ -7,11 +7,11 @@ import copy
 import handtrackingmodule as htm  # mediapipe library used in this module
 
 from Hand import Hand
+from Button import Button
 from ImageCanvas import ImageCanvas
 from Ranking import Ranking
-from states.PaintingState import PaintingState
-from states.MainMenuState import MainMenuState
-from StateMachine import StateMachine
+from State import PaintingState, State, MainMenuState
+from Dataset import Dataset
 
 import math
 
@@ -22,8 +22,7 @@ STATE = "main_menu"
 
 
 # Importing header images using os functions
-# folder_location = # os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "Utilities"))
-folder_location = "../Utilities"
+folder_location = "Utilities"
 headerImage = cv2.imread(f"{folder_location}/Header/header.png", cv2.IMREAD_UNCHANGED)
 ni_logo = cv2.imread(f"{folder_location}/logo.png", cv2.IMREAD_UNCHANGED)
 ni_banner = cv2.imread(f"{folder_location}/banner.png", cv2.IMREAD_UNCHANGED)
@@ -61,22 +60,19 @@ detector = htm.handDetector(detectionCon=0.85)
 # Numpy array with zeros(representing black screen) similar to the dimensions of original video frames
 imageCanvas = ImageCanvas(1280, 720)
 
-state_machine: StateMachine = StateMachine(
-    MainMenuState(
-        headerImage,
-        ni_logo,
-        ni_banner,
-        ranking_img,
-        ranking,
-        video_height,
-        imageCanvas,
-        click_img,
-        erase_img,
-        paint_img,
-        move_img,
-    )
+state: State = MainMenuState(
+    headerImage,
+    ni_logo,
+    ni_banner,
+    ranking_img,
+    ranking,
+    video_height,
+    imageCanvas,
+    click_img,
+    erase_img,
+    paint_img,
+    move_img,
 )
-
 
 hands_list: list[Hand] = []
 
@@ -162,17 +158,17 @@ while True:
         (old_width / new_width, old_height / new_height),
     )
 
-    img = state_machine.run(img, hands_list)
+    state, img = state.run(img, hands_list)
 
-    # for hand in hands_list:
+    #for hand in hands_list:
     #    if hand.middle_finger():
     #        x_min, x_max, y_min, y_max = hand.get_bounding_box()
-    #
+#
     #        y_min = max(0, y_min)
     #        y_max = min(img.shape[0], y_max)
     #        x_min = max(0, x_min)
     #        x_max = min(img.shape[1], x_max)
-    #
+#
     #        #img[y_min:y_max, x_min:x_max] = cv2.GaussianBlur(
     #        #    img[y_min:y_max, x_min:x_max], (77, 77), 77
     #        #)
@@ -183,15 +179,15 @@ while True:
     if key == -1:
         continue
     else:
-        key_consumed = state_machine.handle_input(key)
+        key_consumed, new_state = state.handle_input(key)
 
         if not key_consumed:
             # Global Keyboard Shortcuts
             if key == ord("q"):
                 # Press 'q' to quit the program
                 break
-            # if new_state != None:
-            #     state = new_state
+            if new_state != None:
+                state = new_state
 
 cv2.destroyAllWindows()
 exit()
